PolyDiffusion/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PolyDiffusion"
version = "0.1.0"
description = "Discrete diffusion transformer for guided homopolymer generation"
authors = [{name = "Polygenesis Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "numpy",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["PolyDiffusion*"]
